requirements.txt
Django==4.1.5
django-widget-tweaks==1.4.12
pymysql
tailwind
openpyxl
python-dateutil
mysqlclient



# Example app

django-debug-toolbar
django-bootstrap-form
django-user-sessions

# Documentation

Sphinx
sphinx_rtd_theme

# Build

wheel
twine
